# BackEnd/static/swagger.yaml
openapi: 3.0.0
info:
  title: MoveHive - API
  version: 1.0.0
  description: API para gerenciar usuários e esportes na aplicação MoveHive.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /usuario/RegistrarUsuario:
    post:
      tags:
        - Usuário
      summary: Registrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                NomeCompleto:
                  type: string
                  example: "João da Silva"
                username:
                  type: string
                  example: "joaosilva"
                data_nascimento:
                  type: string
                  format: date
                  example: "2000-05-15"
                email:
                  type: string
                  format: email
                  example: "joao@example.com"
                senha:
                  type: string
                  format: password
                  example: "senhaSegura123"
              required:
                - NomeCompleto
                - username
                - data_nascimento
                - email
                - senha
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Erro na requisição (campos inválidos ou obrigatórios não fornecidos)
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Campos obrigatórios ausentes"
  /usuario/LoginUsuario:
    post:
      summary: Realiza login de usuário
      description: Autentica o usuário através de email e senha. Retorna um token JWT em caso de sucesso.
      tags:
        - Usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: usuario@email.com
                  description: Email do usuário cadastrado
                senha:
                  type: string
                  example: "senha123"
                  description: Senha do usuário
              required:
                - email
                - senha
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    description: Token JWT para autenticação
        '400':
          description: Erro de validação - campos obrigatórios ausentes
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Campos 'email' e 'senha' são obrigatórios"
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Senha incorreta"
        '404':
          description: E-mail não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "E-mail não encontrado"
  /usuario/DadosModal:
    post:
      tags:
        - Usuário
      summary: Adiciona ou atualiza dados do perfil do usuário
      description: Adiciona ou atualiza informações no perfil do usuário autenticado, como foto, biografia, cidade, estado e esportes praticados. Esta rota espera um multipart/form-data.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                foto:
                  type: string
                  format: binary
                  description: "Arquivo de imagem para a foto de perfil do usuário (opcional)."
                dados:
                  type: string
                  description: "Uma string JSON contendo os dados textuais a serem atualizados."
                  example: '{
                    "biografia": "Amo tecnologia e futebol.",
                    "cidade": "São Paulo",
                    "estado": "SP",
                    "esportes_praticados": {
                      "futebol": "iniciante",
                      "basquete": "amador"
                    }
                  }'
      responses:
        '200':
          description: Informações atualizadas com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: sucesso
                  mensagem:
                    type: string
                    example: Informações atualizadas com sucesso
        '400':
          description: Nenhum dado fornecido
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Nenhum dado fornecido
        '401':
          description: Token JWT ausente ou inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Token inválido ou ausente
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Usuário não encontrado
        '500':
          description: Erro interno no servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Erro inesperado no servidor
  /usuario/follow:
    post:
      summary: Seguir um usuário
      description: Permite que um usuário autenticado siga outro usuário, adicionando o ID do usuário seguido na lista de "seguindo" e o ID do seguidor na lista de "seguidores".
      tags:
        - Usuário
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seguido_id:
                  type: string
                  example: "abc123"
                  description: ID do usuário que será seguido.
              required:
                - seguido_id
      responses:
        '200':
          description: Usuário seguido com sucesso.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Usuário seguido com sucesso!"
        '400':
          description: Requisição inválida (ID ausente ou tentativa de seguir a si mesmo).
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Você não pode seguir a si mesmo"
        '404':
          description: Usuário(s) não encontrado(s).
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Usuário(s) não encontrado(s)"
        '500':
          description: Erro interno no servidor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Descrição do erro interno"
  /usuario/unfollow:
    post:
      summary: Deixar de seguir um usuário
      description: Remove a relação de "seguindo" e "seguidores" entre o usuário autenticado e o usuário informado.
      tags:
        - Usuário
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seguido_id:
                  type: string
                  example: "abc123"
                  description: ID do usuário que será deixado de seguir.
              required:
                - seguido_id
      responses:
        '200':
          description: Usuário deixado de seguir com sucesso.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Usuário deixado de seguir com sucesso!"
        '400':
          description: Requisição inválida (ID ausente ou tentativa de deixar de seguir a si mesmo).
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Você não pode deixar de seguir a si mesmo"
        '404':
          description: Usuário(s) não encontrado(s).
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Usuário(s) não encontrado(s)"
        '500':
          description: Erro interno no servidor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Descrição do erro interno"
  /usuario/usuariosSemFiltro:
    get:
      summary: Listar usuários sugeridos (sem filtro)
      description: Retorna todos os usuários que o usuário autenticado ainda não segue.
      tags:
        - Usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuários sugeridos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuarios:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "abc123"
                        username:
                          type: string
                          example: "joao_silva"
                        nome_completo:
                          type: string
                          example: "João da Silva"
                        foto_perfil:
                          type: string
                          example: "https://example.com/foto.jpg"
                        biografia:
                          type: string
                          example: "Apaixonado por futebol e corridas"
                        cidade:
                          type: string
                          example: "São Paulo"
                        estado:
                          type: string
                          example: "SP"
                        esportes_praticados:
                          type: object
                          example:
                            futebol: iniciante
                            corrida: profissional
        '404':
          description: Usuário autenticado não encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Usuário autenticado não encontrado"
        '500':
          description: Erro interno no servidor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Descrição do erro interno"
  /usuario/usuariosComFiltro:
    get:
      summary: Listar usuários sugeridos (com filtro)
      description: >
        Retorna usuários sugeridos com base nos critérios: <br>
        - Cidade igual à do usuário autenticado. <br>
        - Estado igual ao do usuário autenticado. <br>
        - Pelo menos um esporte em comum.   <br> <br>

        Caso nenhum usuário seja encontrado: <br>
        1. Relaxa os critérios, permitindo mais resultados.   <br>
        2. Retorna usuários aleatórios como último recurso. <br>
      tags:
        - Usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuários sugeridos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuarios:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "abc123"
                        username:
                          type: string
                          example: "maria_oliveira"
                        nome_completo:
                          type: string
                          example: "Maria Oliveira"
                        foto_perfil:
                          type: string
                          example: "https://example.com/foto.jpg"
                        biografia:
                          type: string
                          example: "Corredora amadora e apaixonada por vôlei"
                        cidade:
                          type: string
                          example: "Rio de Janeiro"
                        estado:
                          type: string
                          example: "RJ"
                        esportes_praticados:
                          type: object
                          example:
                            futebol: iniciante
                            voleibol: amador
                            corrida: profissional
        '404':
          description: Usuário autenticado não encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Usuário autenticado não encontrado"
        '500':
          description: Erro interno no servidor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Descrição do erro interno"
  /usuario/usuariosSeguidos:
    get:
      summary: Listar usuários seguidos
      description: Retorna todos os usuários que o usuário autenticado segue atualmente.
      tags:
        - Usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuários seguidos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuarios_seguindo:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "abc123"
                        username:
                          type: string
                          example: "joao_silva"
                        nome_completo:
                          type: string
                          example: "João da Silva"
                        foto_perfil:
                          type: string
                          example: "https://example.com/foto.jpg"
                        biografia:
                          type: string
                          example: "Apaixonado por futebol e corridas"
                        cidade:
                          type: string
                          example: "São Paulo"
                        estado:
                          type: string
                          example: "SP"
                        esportes_praticados:
                          type: object
                          example:
                            futebol: amador
                            corrida: amador
        '404':
          description: Usuário autenticado não encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Usuário autenticado não encontrado"
        '500':
          description: Erro interno no servidor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Descrição do erro interno"
  /usuario/usuariosSeguidores:
   get:
      summary: Listar seguidores do usuário autenticado
      description: Retorna todos os usuários que seguem o usuário atualmente autenticado.
      tags:
        - Usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de seguidores.
          content:
            application/json:
              schema:
                type: object
                properties:
                  seguidores:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "def456"
                        username:
                          type: string
                          example: "maria_souza"
                        nome_completo:
                          type: string
                          example: "Maria Souza"
                        foto_perfil:
                          type: string
                          example: "https://example.com/foto_maria.jpg"
                        biografia:
                          type: string
                          example: "Amante de basquete e natação"
                        cidade:
                          type: string
                          example: "Rio de Janeiro"
                        estado:
                          type: string
                          example: "RJ"
                        esportes_praticados:
                          type: object
                          example:
                            basquete: amador
                            natação: profissional
        '404':
          description: Usuário autenticado não encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Usuário autenticado não encontrado"
        '500':
          description: Erro interno no servidor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Descrição do erro interno"
  /usuario/rankingTodos:
    get:
      summary: Lista o ranking de usuários
      description: Retorna a lista de usuários com pontos maiores que zero, ordenados do maior para o menor.
      tags:
        - Usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuários no ranking
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    nome_completo:
                      type: string
                      example: João Silva
                    username:
                      type: string
                      example: joaosilva
                    pontos:
                      type: integer
                      example: 120
                    foto_perfil:
                      type: string
                      format: uri
                      example: https://exemplo.com/imagem.jpg
        '401':
          description: Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Token inválido ou ausente.
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Erro inesperado no servidor.
  /usuario/rankingSeguindo:
    get:
      summary: Lista o ranking dos usuários seguidos
      description: >
        Retorna a lista de usuários que o usuário autenticado está seguindo, incluindo ele próprio,
        ordenados do maior para o menor número de pontos.
      tags:
        - Usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuários no ranking dos seguidos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    nome_completo:
                      type: string
                      example: Maria Oliveira
                    username:
                      type: string
                      example: mariaoliveira
                    pontos:
                      type: integer
                      example: 95
                    foto_perfil:
                      type: string
                      format: uri
                      example: https://exemplo.com/imagem_maria.jpg
        '401':
          description: Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Token inválido ou ausente.
        '404':
          description: Usuário autenticado não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Usuário autenticado não encontrado no banco de dados
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Erro inesperado no servidor.
  /usuario/esqueciSenha:
    post:
      tags:
        - Usuário
      summary: Solicitar reset de senha
      description: Permite que um usuário solicite a redefinição de senha fornecendo seu e-mail.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@exemplo.com
            examples:
              emailExample:
                summary: Exemplo de requisição
                value:
                  email: usuario@exemplo.com
      responses:
        '200':
          description: Solicitação de reset de senha realizada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: Um e-mail de redefinição de senha foi enviado.
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: Formato de e-mail inválido ou ausente.
  /usuario/verificarCodigo:
     post:
      tags:
        - Usuário
      summary: Verificar código de reset de senha
      description: Verifica se o código de reset enviado pelo usuário é válido e não expirou.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - codigo
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@exemplo.com
                codigo:
                  type: string
                  example: "123456"
            examples:
              exemploRequisicao:
                summary: Exemplo de requisição
                value:
                  email: usuario@exemplo.com
                  codigo: "123456"
      responses:
        '200':
          description: Código válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  valido:
                    type: boolean
                    example: true
                  msg:
                    type: string
                    example: Código verificado com sucesso.
        '400':
          description: Código inválido ou expirado
          content:
            application/json:
              schema:
                type: object
                properties:
                  valido:
                    type: boolean
                    example: false
                  msg:
                    type: string
                    example: Código inválido ou expirado.
...